test(monorepo): use relative paths in snapshots (same as in core/cli)
